refactor: change SearchBar.tsx
